9043/refactor: extract Point calculation logic into PointCalculatorService from PaymentCalculatorService
